Adding report_email.py to the repo.
It is part of Project 4: Automating Uploading Catalog Info.

The script calls on reports.py and emails.py. It is the main script for Project 4.

Its functions are: create_dict, create_paragraph and main.

It opens product catalog description text files, creates a pdf report detailing
descriptions and images uploaded to the Catalog and emails the pdf to the user.