Adding health_check.py to the repo.
This script is part of Project 4: Automating Uploading Catalog Info.

It checks the health of the pc and sends an email alert when something is
wrong.

It calls on emails.py to generate and send the email alert.

The script is meant to be run periodically as a cron job.